Regression for groovy routes #118
